Project maintenance and review
This commit is a collection of changes related to:

- Code style fixes (including nolint) and formatting.
- Suppression of compiler warnings in bitwise operations.
- Minor edits (e.g. copyright dates).

+++ register/Internals.h
@@ -2,7 +2,7 @@
 /**
  * @file      Internals.h
  * @author    Nicolas Clauvelin (nclauvelin@sendyne.com)
- * @copyright Copyright 2010-2019 Sendyne Corp. All rights reserved.
+ * @copyright Copyright 2010-2022 Sendyne Corp. All rights reserved.
  *
  * This header collects various implementations which are required for cppreg
  * implementation but not intended to be fully exposed to the user.
@@ -31,7 +31,8 @@ namespace internals {
  * There is overflow if value if strictly larger than limit.
  */
 template <typename T, T value, T limit>
-struct check_overflow : std::integral_constant<bool, value <= limit> {};
+struct check_overflow    // NOLINT
+    : std::integral_constant<bool, value <= limit> {};
 
 
 //! is_aligned implementation.
@@ -42,8 +43,10 @@ struct check_overflow : std::integral_constant<bool, value <= limit> {};
  * This will only derived from std::true_type if the address is aligned.
  */
 template <Address address, std::size_t alignment>
-struct is_aligned
-    : std::integral_constant<bool, (address & (alignment - 1)) == 0> {};
+struct is_aligned    // NOLINT
+    : std::integral_constant<
+          bool,
+          (static_cast<std::size_t>(address) & (alignment - 1)) == 0> {};
 
 
 }    // namespace internals
